+++ CHANGES.rst
@@ -1,5 +1,13 @@
 
 
+1.0.4 (27 May 2020)
+-------------------
+
+ - Add a script (pypeit_flux_setup) for creating fluxing,
+   coadd1d and tellfit pypeit files
+ - Add telluric fitting script, pypeit_tellfit
+
+
 1.0.3 (04 May 2020)
 -------------------